definitvely fixed crashes and bugs
This version is the most stable up-to-date. Crashes has been fixed.
Bug that a msg is not received sometimes has been fixed. (prio problem).
the FOTA hasn't been tested yet so the release will occur on monday
instead.

+++ Shrpe.h
@@ -48,7 +48,7 @@ class Shrpe
 	  @param	size		size of the array.
 	  @return				the length of the array.
 	*/
-	byte read(uint8_t* buffer_ptr, uint8_t size);
+	int read(uint8_t* buffer_ptr, uint8_t size);
 	/*
 	  @Return	true if the shield has data available, otherwise false.
 	*/
